udpate names and cleanup
