+++ Compiler/build.gradle
@@ -24,21 +24,10 @@ test {
     useJUnitPlatform()
 }
 
-
-task testLO1(type: Test) {
-    useJUnitPlatform()
-    filter {
-        includeTestsMatching "assignment2.LiveOak1CompilerTest"
-    }
-    testLogging {
-        showStandardStreams = true
-    }
-}
-
 task testLO3(type: Test) {
     useJUnitPlatform()
     filter {
-        includeTestsMatching "assignment2.LiveOak3CompilerTest"
+        includeTestsMatching "compiler.CompilerTest"
     }
     testLogging {
         showStandardStreams = true
@@ -52,7 +41,7 @@ java {
 
 jar {
     manifest {
-        attributes 'Main-Class': 'assignment2.LiveOak3Compiler'
+        attributes 'Main-Class': 'compiler.Compiler'
     }
     sourceCompatibility = JavaVersion.VERSION_11
     targetCompatibility = JavaVersion.VERSION_11
