--- Compiler/src/main/java/assignment2/enums/BinopType.java
@@ -1,7 +1,0 @@
-package assignment2;
-
-public enum BinopType {
-    ARITHMETIC,
-    BITWISE,
-    COMPARISON,
-}
