--- Compiler/src/main/java/assignment2/enums/LabelType.java
@@ -1,7 +1,0 @@
-package assignment2;
-
-public enum LabelType {
-    DEFAULT,
-    RETURN,
-    BREAK,
-}
