--- Compiler/src/main/java/assignment2/enums/Statement.java
@@ -1,10 +1,0 @@
-package assignment2;
-
-public enum Statement {
-    NULL,
-    RETURN,
-    BREAK,
-    LOOP,
-    CONDITIONAL,
-    ASSIGN,
-}
