--- Compiler/src/main/java/assignment2/errors/SyntaxErrorException.java
@@ -1,12 +1,0 @@
-package assignment2.errors;
-
-public class SyntaxErrorException extends CompilerException {
-    public SyntaxErrorException(String message, int line) {
-        super(message, line);
-    }
-
-    @Override
-    public String toString() {
-        return "Syntax error at line " + getLine() + ": " + getMessage();
-    }
-}
