--- Compiler/src/main/java/assignment2/errors/TypeErrorException.java
@@ -1,12 +1,0 @@
-package assignment2.errors;
-
-public class TypeErrorException extends CompilerException {
-    public TypeErrorException(String message, int line) {
-        super(message, line);
-    }
-
-    @Override
-    public String toString() {
-        return "Type error at line " + getLine() + ": " + getMessage();
-    }
-}
