+++ Compiler/src/main/java/compiler/enums/BinopType.java
@@ -1,0 +1,7 @@
+package compiler;
+
+public enum BinopType {
+    ARITHMETIC,
+    BITWISE,
+    COMPARISON,
+}
