+++ Compiler/src/main/java/compiler/enums/LabelType.java
@@ -1,0 +1,7 @@
+package compiler;
+
+public enum LabelType {
+    DEFAULT,
+    RETURN,
+    BREAK,
+}
