+++ Compiler/src/main/java/compiler/enums/Statement.java
@@ -1,0 +1,10 @@
+package compiler;
+
+public enum Statement {
+    NULL,
+    RETURN,
+    BREAK,
+    LOOP,
+    CONDITIONAL,
+    ASSIGN,
+}
