+++ Compiler/src/main/java/compiler/errors/TypeErrorException.java
@@ -1,0 +1,12 @@
+package compiler.errors;
+
+public class TypeErrorException extends CompilerException {
+    public TypeErrorException(String message, int line) {
+        super(message, line);
+    }
+
+    @Override
+    public String toString() {
+        return "Type error at line " + getLine() + ": " + getMessage();
+    }
+}
