--- Compiler/src/test/java/assignment2/LiveOak1CompilerTest.java
@@ -1,107 +1,0 @@
-package assignment2;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-import assignment2.errors.TypeErrorException;
-import java.io.ByteArrayOutputStream;
-import java.io.FileDescriptor;
-import java.io.FileOutputStream;
-import java.io.PrintStream;
-import java.nio.file.Path;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-/*******************
- * Example test cases for main.LiveOak3Compiler, on LiveOak-1
- *
- *
- * *******************/
-class LiveOak1CompilerTest {
-
-    private static final String lo1ValidProgramDir = Path.of(
-        "src",
-        "test",
-        "resources",
-        "LO-1",
-        "ValidPrograms"
-    ).toString();
-    private static final String lo1InvalidProgramDir = Path.of(
-        "src",
-        "test",
-        "resources",
-        "LO-1",
-        "InvalidPrograms"
-    ).toString();
-
-    private static final String lo2ValidProgramDir = Path.of(
-        "src",
-        "test",
-        "resources",
-        "LO-2",
-        "ValidPrograms"
-    ).toString();
-    private static final String lo2InvalidProgramDir = Path.of(
-        "src",
-        "test",
-        "resources",
-        "LO-2",
-        "InvalidPrograms"
-    ).toString();
-
-    private static final String lo3GoodExampleDir = Path.of(
-        "src",
-        "test",
-        "resources",
-        "LO-3",
-        "ValidPrograms"
-    ).toString();
-    private static final String lo3BadExampleDir = Path.of(
-        "src",
-        "test",
-        "resources",
-        "LO-3",
-        "InvalidPrograms"
-    ).toString();
-    private static ByteArrayOutputStream errContent;
-
-    @BeforeEach
-    void setUp() {
-        errContent = new ByteArrayOutputStream();
-        System.setErr(new PrintStream(errContent));
-    }
-
-    @AfterEach
-    void tearDown() {
-        System.setOut(
-            new PrintStream(new FileOutputStream(FileDescriptor.out))
-        );
-    }
-
-    private static String getStdErr() {
-        return errContent.toString().replaceAll("\r", "");
-    }
-
-    /** VALID PROGRAMS
-     **/
-    @Test
-    @DisplayName("Test_0")
-    void testL01_0() throws Throwable {
-        String fileName = Path.of(lo3GoodExampleDir, "test_45.lo").toString();
-        String program = LiveOak3Compiler.compiler(fileName);
-        SamTestRunner.checkReturnValue(program, 296);
-        // System.out.println(program);
-        // assertThrows(
-        //     Error.class,
-        //     () -> LiveOak3Compiler.compiler(fileName),
-        //     "Expected parse error to be thrown for file test_9.lo"
-        // );
-        // assertTrue(
-        //     getStdErr()
-        //         .contains(
-        //             "Failed to compile src/test/resources/LO-2/InvalidPrograms/test_9.lo"
-        //         )
-        // );
-    }
-}
